+++ src/django_admin_select_filter/__init__.py
@@ -2,6 +2,7 @@ from django_admin_select_filter.filters import (
     BaseSelectFilter,
     ChoiceFilter,
     ForeignKeyFilter,
+    field_list_filter,
 )
 from django_admin_select_filter.urls import django_admin_select_filter_path
 from django_admin_select_filter.views import Select2FilterOptionsView
@@ -14,4 +15,5 @@ __all__ = [
     "ForeignKeyFilter",
     "Select2FilterOptionsView",
     "django_admin_select_filter_path",
+    "field_list_filter",
 ]